requirements.txt
dnspython>=2.4.2
requests>=2.31.0
netifaces>=0.11.0
speedtest-cli>=2.1.3
stun>=0.1.12
